createpage bug fix: price_normalized to float

+++ backend/recipick/models.py
@@ -16,7 +16,7 @@ class Ingredient(models.Model):
     quantity = models.FloatField()                    # 구매 단위!! normalized price할 때 주로 사용되고, 유저에게 한 
                                                       # ingredient를 살 때 얼마 지불해야하는지 알리는 가격 (NOT quantity of ingredient the recipe needs)
     price = models.IntegerField()
-    price_normalized = models.IntegerField(null=True) # 좀 더 엄밀한 나눗셈 필요
+    price_normalized = models.FloatField(null=True) # 좀 더 엄밀한 나눗셈 필요
     igd_type = models.CharField(max_length=5) # 단위: 개 / g
     brand = models.CharField(max_length=64)
     picture = models.ImageField(null=True)
